project/microservices/language/requirements.txt
faust-streaming
faust-streaming-rocksdb
langdetect
